requirements.txt
streamlit==1.24.0
streamlit-authenticator==0.2.2
pandas==2.0.3
plotly==5.15.0
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0
langchain==0.0.200
openai==0.27.8
python-jose==3.3.0
sqlalchemy==2.0.19
